requirements.txt
defusedxml==0.4.1
Django==1.9.6
django-activity-stream==0.6.2
django-allauth==0.25.2
oauthlib==1.1.1
python3-openid==3.0.10
pytz==2016.4
requests==2.9.1
requests-oauthlib==0.6.1
djangorestframework==3.3.3
